[STF] clang-tidy batch 3: the 58 findings of the second run
- bugprone-exception-escape (26): all reported at on_throw's operator<<,
  the noexcept runner, for its instantiations with test lambdas that throw.
  In clang-tidy's device pass _CCCL_TRY expands to no catch, so the check
  sees the callable's throw escape the runner. One NOLINT with the reason on
  the runner; the per-block suppressions in the tests, which addressed the
  wrong location, are removed.
- The runner's callable binding is back to trunk's non-const reference (a
  mutable callable, re-invoked by retry, needs it) and suppressed with the
  existing comment as the reason.
- const restored on default-initialized contexts/handles/hashtables (they
  compile; only event_list's defaulted constructor does not, suppressed).
- CTAD tests assert the const type; remaining narrowing, string-find,
  reserve() and branch-clone sites fixed or suppressed with a reason.

+++ cudax/examples/stf/graph_algorithms/pagerank_batched.cu
@@ -47,7 +47,7 @@ __device__ void calculating_pagerank(
     const int out_degree = loffsets[neighbor + 1] - loffsets[neighbor];
     rank_sum += lpage_rank[neighbor] / static_cast<float>(out_degree);
   }
-  lnew_page_rank[idx] = static_cast<float>(0.85) * rank_sum + (1.0 - 0.85) * lpersonalization[idx];
+  lnew_page_rank[idx] = static_cast<float>(0.85) * rank_sum + static_cast<float>(1.0 - 0.85) * lpersonalization[idx];
 }
 
 /**
